+++ src/FixtureLoader.php
@@ -17,8 +17,7 @@ class FixtureLoader
     public function __construct(
         \Traversable $fixtures,
         private readonly FixtureHelper $helper,
-    )
-    {
+    ) {
         $this->fixtures = iterator_to_array($fixtures);
     }